+++ lib/receipts.rb
@@ -5,9 +5,16 @@ require "prawn/table"
 
 module Receipts
   autoload :Base, "receipts/base"
+  autoload :CertificateOfOrigin, "receipts/certificate_of_origin"
+  autoload :CommercialInvoice, "receipts/commercial_invoice"
   autoload :Invoice, "receipts/invoice"
+  autoload :PackingList, "receipts/packing_list"
   autoload :Receipt, "receipts/receipt"
   autoload :Statement, "receipts/statement"
+
+  module Concerns
+    autoload :DocumentComponents, "receipts/concerns/document_components"
+  end
 
   @@default_font = nil